requirements.txt
Flask
git+https://github.com/openai/whisper.git
google-generativeai
python-dotenv
openai
